refactor: limpeza no código
