+++ main.cpp
@@ -1,398 +1,17 @@
 #include <iostream>
 #include <string>
-#include <fstream>
 #include <vector>
 
-#include "utils/EveryLetterIsLower.cpp"
+#include "utils/IsLower.cpp"
+#include "utils/readFile.cpp"
+#include "functions/fatos.cpp"
+#include "functions/lexico.cpp"
+#include "functions/sintaticoRegras.cpp"
 
 using namespace std;
 
-struct variavel{
-    string nome;
-    string valor;
-};
-
-struct regra{
-    vector<pair<variavel, string>> se;
-    vector<pair<variavel, string>> entao;
-};
-
+vector<variavel> MT;
 vector<regra> regras;
-regra regraAtual;
-
-vector<variavel> variavelMT;
-
-vector<string> readFile(string filename)
-{
-    string line;
-    vector<string> lines;
-    ifstream fileToRead(filename);
-    if (fileToRead.is_open())
-    {
-        while (getline(fileToRead, line, '\n'))
-        {
-            string newline = "";
-            for (char character: line)
-            {
-                if (!isspace(character))
-                {
-                    newline.push_back(character);
-                }
-            }
-            lines.push_back(newline);
-        }
-        fileToRead.close();
-        return lines;
-    }
-    throw runtime_error("Could not read file");
-}
-
-string substring(string line, int inicio, int final)
-{
-    string subs = "";
-    for (int i = inicio; i < final; i++)
-    {
-        subs.push_back(line[i]);
-    }
-    return subs;
-}
-
-string s2(string line, int *characterStopped)
-{
-    for (int i = *characterStopped; i < line.length(); i++)
-    {
-        if (isalpha(line[i]) && isupper(line[i]))
-        {
-            string subs = substring(line, *characterStopped, i);
-            if (subs.compare("true") == 0 || subs.compare("false") == 0)
-            {
-                //cout << subs << " is a keyword" << endl;
-                *characterStopped = i;
-            }
-            else
-            {
-                // erro
-                *characterStopped = -1;
-            }
-            return subs;
-        }
-        if (!isalnum(line[i]))
-        {
-            string subs = substring(line, *characterStopped, i);
-            if (subs.compare("true") == 0 || subs.compare("false") == 0)
-            {
-                //cout << subs << " is a keyword" << endl;
-            }
-            else
-            {
-                //cout << subs << " is a identifier" << endl;
-            }
-            
-            *characterStopped = i;
-            return subs;
-        }
-    }
-
-    string subs = substring(line, *characterStopped, line.length());
-    if (subs.compare("true") == 0 || subs.compare("false") == 0)
-    {
-        //cout << subs << " is a keyword" << endl;
-    }
-    else
-    {
-        //cout << subs << " is a identifier" << endl;
-    }
-    *characterStopped = -1;
-    return subs;
-}
-
-string s3(string line, int *characterStopped)
-{
-    for (int i = *characterStopped; i < line.length(); i++)
-    {
-        if (isalpha(line[i]))
-        {
-            if (islower(line[i]))
-            {
-                // print
-                string subs = substring(line, *characterStopped, i);
-                //cout << subs << " is a keyword" << endl;
-                *characterStopped = i;
-                return subs;
-            }
-        }
-        else
-        {
-            // print
-            string subs = substring(line, *characterStopped, i);
-            //cout << subs << " is a keyword" << endl;
-            *characterStopped = i;
-            return subs;
-        }
-    }
-
-    // print
-    string subs = substring(line, *characterStopped, line.length());
-    //cout << subs << " is a keyword" << endl;
-    *characterStopped = -1;
-    return subs;
-}
-
-string s1(string line, int *characterStopped)
-{
-    if (islower(line[*characterStopped]))
-    {
-        return s2(line, characterStopped);
-    }
-    else
-    {
-        return s3(line, characterStopped);
-    }
-}
-
-string s4(string line, int *characterStopped)
-{
-    string op (1, line[*characterStopped]);
-
-    *characterStopped += 1;
-    //cout << op << " is a operator" << endl;;
-
-    if (*characterStopped >= line.length())
-    {
-        *characterStopped = -1;
-        if (line[*characterStopped] == '&' || line[*characterStopped] == '=')   return op;
-        throw runtime_error("caractere nao reconhecido " + line[*characterStopped]);
-    }
-
-    if (isalpha(line[*characterStopped]))
-    {
-        if (islower(line[*characterStopped]))
-        {
-            return op;
-        }
-    }
-    throw runtime_error("error de sintaxe, esperando letra minuscula depois de operador\nlinha: " + line + "caractere" + line[*characterStopped]);
-}
-
-string s0(string line, int *characterStopped)
-{
-    if (isalpha(line[*characterStopped]))
-    {
-        return s1(line, characterStopped);
-    }
-    else
-    {
-        return s4(line, characterStopped);
-    }
-}
-
-/*
-S8
-*/
-void S8(vector<string> rule, int *string_stopped)
-{
-    if (*string_stopped >= rule.size())
-    {
-        regras.push_back(regraAtual);
-
-        regraAtual.se.clear();
-        regraAtual.entao.clear();
-
-        *string_stopped = -1;
-        return;
-    }
-    if (rule[*string_stopped].compare("&") == 0)
-    {
-        return;
-    }
-    throw runtime_error("Esperando &");
-}
-
-/*
-S7
-*/
-void S7(vector<string> rule, int *string_stopped, int linha, variavel varENTAO, string operador)
-{
-    if (*string_stopped >= rule.size())
-    {
-        throw runtime_error("Esperando true/false");
-    }
-    if (rule[*string_stopped].compare("true") == 0 || rule[*string_stopped].compare("false") == 0)
-    {
-        varENTAO.valor = rule[*string_stopped];
-
-        regraAtual.entao.push_back({varENTAO, operador});
-        //regras[linha].entao.push_back(varENTAO);
-
-        //variavelENTAO.push_back(varENTAO);
-
-        *string_stopped += 1;
-        S8(rule, string_stopped);
-        return;
-    }
-    throw runtime_error("Esperando true/false");
-}
-
-/*
-S6
-*/
-void S6(vector<string> rule, int *string_stopped, int linha, variavel varENTAO, string operador)
-{
-    if (*string_stopped >= rule.size())
-    {
-        throw runtime_error("Esperando um '='");
-    }
-    if (rule[*string_stopped].compare("=") == 0)
-    {
-        *string_stopped += 1;
-        S7(rule, string_stopped, linha, varENTAO, operador);
-        return;
-    }
-    throw runtime_error("Esperando um '='");
-}
-
-/*
-S5
-*/
-void S5(vector<string> rule, int *string_stopped, int linha, string operador)
-{
-    if (*string_stopped >= rule.size())
-    {
-        throw runtime_error("Esperando variaveis");
-    }
-    if (EveryLetterIsLower(rule[*string_stopped]))
-    {   
-        variavel varENTAO;
-        varENTAO.nome = rule[*string_stopped];
-
-        *string_stopped += 1;
-        S6(rule, string_stopped, linha, varENTAO, operador);
-        return;
-    }
-    throw runtime_error("Todas as letras de uma variável devem ser minúsculas");
-}
-
-/*
-S4
-*/
-void S4(vector<string> rule, int *string_stopped)
-{
-    if (*string_stopped >= rule.size())
-    {
-        throw runtime_error("Esperando um 'ENTAO' ou '&'");
-    }
-    if (rule[*string_stopped].compare("&") == 0 || rule[*string_stopped].compare("ENTAO") == 0)
-    {
-        return;
-    }
-}
-
-/*
-S3
-*/
-void S3(vector<string> rule, int *string_stopped, int linha, variavel varSE, string operador)
-{
-    if (*string_stopped >= rule.size())
-    {
-        throw runtime_error("Esperando true/false");
-    }
-    if (rule[*string_stopped].compare("true") == 0 || rule[*string_stopped].compare("false") == 0)
-    {
-        varSE.valor = rule[*string_stopped];
-        //varSE.linha = linha;
-
-        regraAtual.se.push_back({varSE, operador});
-        //regras[linha].se.push_back(varSE);
-
-        //variavelSE.push_back(varSE);
-
-        *string_stopped += 1;
-        S4(rule, string_stopped);
-        return;
-    }
-    throw runtime_error("Esperando true/false");
-}
-
-/*
-S2
-*/
-void S2(vector<string> rule, int *string_stopped, int linha, variavel varSE, string operador)
-{
-    if (*string_stopped >= rule.size())
-    {
-        throw runtime_error("Esperando um '='");
-    }
-    if (rule[*string_stopped].compare("=") == 0)
-    {
-        *string_stopped += 1;
-        S3(rule, string_stopped, linha, varSE, operador);
-        return;
-    }
-    throw runtime_error("Esperando um '=' depois da variavel");
-}
-
-/*
-S1
-*/
-void S1(vector<string> rule, int *string_stopped, int linha, string operador)
-{
-    if (*string_stopped >= rule.size())
-    {
-        throw runtime_error("Esperando uma expressao depois do 'SE'");
-    }
-    if (EveryLetterIsLower(rule[*string_stopped]))
-    {
-        variavel varSE;
-        varSE.nome = rule[*string_stopped];
-
-        *string_stopped += 1;
-        S2(rule, string_stopped, linha, varSE, operador);
-        return;
-    }
-    throw runtime_error("Todas as letras de uma variável devem ser minúsculas");
-}
-
-/*
-S0
-*/
-void S0(vector<string> rule, int *string_stopped, bool *entao, int linha)
-{
-    if (rule[*string_stopped].compare("SE") == 0 && *string_stopped == 0)
-    {
-        *string_stopped += 1;
-        S1(rule, string_stopped, linha, "");
-        return;
-    }
-    else if (*string_stopped != 0)
-    {
-        //if (rule[*string_stopped].compare("&") == 0)
-        if (rule[*string_stopped].compare("&") == 0 || rule[*string_stopped].compare("|") == 0)
-        {
-            *string_stopped += 1;
-            if (*entao)
-            {
-                S5(rule, string_stopped, linha, rule[*string_stopped-1]);
-            }
-            else
-            {
-                S1(rule, string_stopped, linha, rule[*string_stopped-1]);
-            }
-            return;
-        }
-        else if (rule[*string_stopped].compare("ENTAO") == 0)
-        {
-            if (*entao) // 2 entao na msm linha
-            {
-                throw runtime_error("2 'ENTAO' na mesma linha");
-            }
-            *string_stopped += 1;
-            *entao = true;
-            S5(rule, string_stopped, linha, "");
-            return;
-        }
-    }
-    throw runtime_error("error de sintaxe, esperando 'SE'");
-}
 
 vector<vector<string>> analyzer(vector<string> lines)
 {
@@ -416,11 +35,11 @@ vector<vector<string>> analyzer(vector<string> lines)
 bool encadeamentoParaTras(variavel objetivo)
 {
     // verificar minha MT
-    for (int i = 0; i < variavelMT.size(); i++)
+    for (int i = 0; i < MT.size(); i++)
     {
-        if (variavelMT[i].nome.compare(objetivo.nome) == 0)
+        if (MT[i].nome.compare(objetivo.nome) == 0)
         {
-            if (variavelMT[i].valor.compare(objetivo.valor) == 0)
+            if (MT[i].valor.compare(objetivo.valor) == 0)
             {
                 return true;
             }
@@ -496,68 +115,11 @@ bool encadeamentoParaTras(variavel objetivo)
             varMT.valor = objetivo.valor;
 
             cout << objetivo.nome << "=" << objetivo.valor << endl;
-            variavelMT.push_back(varMT);
+            MT.push_back(varMT);
             return true;
         }
     }
     return false;
-}
-
-void f3(vector<string> fact, int pos, variavel varMT)
-{
-    if (pos < fact.size())
-    {
-        throw runtime_error("erro na base de fatos");
-    }
-    //varMT.linha = -1;
-    variavelMT.push_back(varMT);
-}
-
-void f2(vector<string> fact, int pos, variavel varMT)
-{
-    if (pos >= fact.size())
-    {
-        throw runtime_error("erro na base de fatos");
-    }
-    if (fact[pos].compare("true") == 0 || fact[pos].compare("false") == 0)
-    {
-        varMT.valor = fact[pos];
-        f3(fact, pos+1, varMT);
-    }
-    else
-    {
-        throw runtime_error("erro na base de fatos");
-    }
-}
-
-void f1(vector<string> fact, int pos, variavel varMT)
-{
-    if (pos >= fact.size())
-    {
-        throw runtime_error("erro na base de fatos");
-    }
-    if (fact[pos].compare("=") == 0)
-    {
-        f2(fact, pos+1, varMT);
-    }
-    else
-    {
-        throw runtime_error("erro na base de fatos");
-    }
-}
-
-void f0(vector<string> fact)
-{
-    if (EveryLetterIsLower(fact[0]))
-    {
-        variavel varMT;
-        varMT.nome = fact[0];
-        f1(fact, 1, varMT);
-    }
-    else
-    {
-        throw runtime_error("erro na base de fatos");
-    }
 }
 
 int main()
@@ -571,17 +133,6 @@ int main()
     
     vector<vector<string>> rules = analyzer(lines);
 
-    cout << endl << "regras: " << endl;
-    for (int i = 0; i < rules.size(); i++)
-    {
-        cout << "Rule " << i + 1 << ": ";
-        for (string token : rules[i])
-        {
-            cout << token << " ";
-        }
-        cout << endl;
-    }
-
     for (int i = 0; i < rules.size(); i++)
     {
         if (rules[i].size() == 0) continue;
@@ -592,6 +143,7 @@ int main()
             //cout << "pos " << string_stopped << "string: " << rule[string_stopped] << endl;
             S0(rules[i], &string_stopped, &imInEntao, i);
         }
+        regras.push_back(atual());
     }
 
     cout << endl << endl;
@@ -622,14 +174,14 @@ int main()
 
     for (vector<string> r : slavei)
     {
-        f0(r);
+        MT.push_back(f0(r));
     }
 
     cout << "MT:" << endl;
 
-    for (int i = 0; i < variavelMT.size(); i++)
+    for (int i = 0; i < MT.size(); i++)
     {
-        cout << variavelMT[i].nome << "=" << variavelMT[i].valor << endl;
+        cout << MT[i].nome << "=" << MT[i].valor << endl;
     }
 
     cout << endl;
@@ -643,6 +195,12 @@ int main()
         cout << "TRUE";
     } else {
         cout << "FALSE";
+    }
+
+    cout << endl << "MT ao final:" << endl;
+    for (auto s : MT)
+    {
+        cout << s.nome << " = " << s.valor << endl;
     }
 
     // testar o false para ver se tem conflito
