+++ utils/readFile.cpp
@@ -1,0 +1,31 @@
+#include <iostream>
+#include <fstream>
+#include <vector>
+#include <string>
+
+using namespace std;
+
+vector<string> readFile(string filename)
+{
+    string line;
+    vector<string> lines;
+    ifstream fileToRead(filename);
+    if (fileToRead.is_open())
+    {
+        while (getline(fileToRead, line, '\n'))
+        {
+            string newline = "";
+            for (char character: line)
+            {
+                if (!isspace(character))
+                {
+                    newline.push_back(character);
+                }
+            }
+            lines.push_back(newline);
+        }
+        fileToRead.close();
+        return lines;
+    }
+    throw runtime_error("Could not read file");
+}
