+++ utils/substring.cpp
@@ -1,0 +1,14 @@
+#include <string>
+#include <vector>
+
+using namespace std;
+
+string substring(string line, int inicio, int final)
+{
+    string subs = "";
+    for (int i = inicio; i < final; i++)
+    {
+        subs.push_back(line[i]);
+    }
+    return subs;
+}
